requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.2.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
coverage==5.3
DateTime==4.3
distlib==0.3.1
dnspython==2.0.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
identify==1.5.5
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.9.1
mccabe==0.6.1
more-itertools==8.5.0
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pathspec==0.8.0
pbr==5.5.0
pluggy==0.13.1
pre-commit==2.7.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
SQLAlchemy==1.3.19
stevedore==3.2.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
